Pull please

+++ app/src/main/java/com/example/abhinav/buzzer/HomeSingleActivity.java
@@ -95,56 +95,36 @@ public class HomeSingleActivity extends AppCompatActivity {
         mHomeSingleRemoveBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-
-                mDatabase.child(mPost_key).removeValue();
-//                final DatabaseReference College=FirebaseDatabase.getInstance().getReference("College").child(getIntent().getStringExtra("CollegeId"));
-//                College.runTransaction(new Transaction.Handler() {
-//                    @Override
-//                    public Transaction.Result doTransaction(MutableData mutableData) {
-//                        CollegeName artist=mutableData.getValue(CollegeName.class);
-//                        if(artist==null)
-//                        {
-//                            return Transaction.success(mutableData);
-//                        }
-//                        else
-//                        {
-//                        int startTd=artist.getFirstPost();
-//                        int endId=artist.getLastPost();
-//                        if(startTd==endId)
-//                        {
-//                            mutableData.child("FirstPost").setValue(null);
-//                            mutableData.child("LastPost").setValue(null);
-//
-//                        }
-//                        else
-//                        {
-//                            endId--;
-//                            mutableData.child("LastPost").setValue(endId);
-//                        }
-//                    }
-//                    return Transaction.success(mutableData);
-//                    }
-//
-//                    @Override
-//                    public void onComplete(DatabaseError databaseError, boolean b, DataSnapshot dataSnapshot) {
-//
-//                    }
-//                });
-                mStorage= FirebaseStorage.getInstance().getReference().child("Posts/"+postId);
-                mStorage.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                mDatabase.addListenerForSingleValueEvent(new ValueEventListener() {
                     @Override
-                    public void onSuccess(Void aVoid) {
-                        Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();                    }
+                    public void onDataChange(DataSnapshot dataSnapshot) {
+                      Long val=(Long)dataSnapshot.child(mPost_key).child("With_image").getValue();
+                        if(val==1)
+                        {
+                            mStorage= FirebaseStorage.getInstance().getReference().child("Posts/"+postId);
+                            mStorage.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                                @Override
+                                public void onSuccess(Void aVoid) {
+                                    Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();                    }
 
-                }).addOnFailureListener(new OnFailureListener() {
+                            }).addOnFailureListener(new OnFailureListener() {
+                                @Override
+                                public void onFailure(@NonNull Exception exception) {
+                                    Toast.makeText(HomeSingleActivity.this,"Failed",Toast.LENGTH_SHORT).show();
+                                }
+                            });
+                        }
+                        else
+                            Toast.makeText(HomeSingleActivity.this,"Removed Successfully",Toast.LENGTH_SHORT).show();
+                    }
+
                     @Override
-                    public void onFailure(@NonNull Exception exception) {
-                        Toast.makeText(HomeSingleActivity.this,"Failed",Toast.LENGTH_SHORT).show();
+                    public void onCancelled(DatabaseError databaseError) {
+
                     }
                 });
-                //}
-               // else
-              //      Toast.makeText(HomeSingleActivity.this,"Removed Succesfully",Toast.LENGTH_SHORT).show();
+
+                mDatabase.child(mPost_key).removeValue();
 
                 Intent mainIntent = new Intent(HomeSingleActivity.this, MainActivity.class);
                 startActivity(mainIntent);
